organize env vars
