+++ scripts/.env.defaults
@@ -1,0 +1,6 @@
+# Order of precedence is Default -> Private.
+# Any default/public vars can be overwritten in private env file
+
+# default private key for optimism
+PRIVATE_KEY=ac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
+INFURA_KEY=561ebceae957407ea6699a474aa4f7b0
